Separated Amount / SingleAmount / PostingAmount into separated moduiles.
Those are similar, but having in the same file clutters the structure.
Separated into different files also revealed low coverage for
SingleAmount / PostingAmount methods.
